+++ src/main/java/com/ddabong/TripFlow/config/jasypt/EncryptConfig.java
@@ -1,12 +1,14 @@
-package com.ddabong.TripFlow.config;
+package com.ddabong.TripFlow.config.jasypt;
 
 import jakarta.annotation.PostConstruct;
+import org.jasypt.encryption.pbe.PooledPBEStringEncryptor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class EncryptConfig {
 
+    /* datasource
     @Value("${spring.datasource.url}")
     private String url;
 
@@ -15,6 +17,7 @@ public class EncryptConfig {
 
     @Value("${spring.datasource.password}")
     private String password;
+     */
 
     @PostConstruct
     public void init() {
